refactor(cli): extract doctor, toolchain, and rust_plan modules (#339)
Continue main.rs decomposition: lift the `soldr doctor` implementation,
the toolchain passthroughs / `soldr rustup` / `soldr toolchain
install|prepare`, and the Rust artifact plan + warm-restore +
thin-slice manifest machinery into dedicated sibling modules.

Refs #339
